got the single gpu pipeline working
